1402-1403-second-term/cloud/HW2/phase1/requirements.txt
Flask~=2.3.3
redis~=5.0.1
elasticsearch~=7.9.0
PyMovieDb~=0.0.9
requests~=2.31.0
python-dotenv~=1.0.1
lxml_html_clean